test: Refactor test codes to add age input logic
